Estilos a EmojiButton
