+++ src/EmojiButton/EmojiButton.css
@@ -1,0 +1,23 @@
+.EmojiButton {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    margin-top: 40px;
+    font-size: 20px;
+    font-weight: 600;
+}
+
+.EmojiButton select {
+    width: 100px;
+    font-size: 20px;
+    padding: 5px 10px;
+    margin-top: 10px;
+    text-align: center;
+    border: 3px solid #dedede;
+    border-radius: 10px;
+}
+
+.EmojiButton select:focus {
+    outline: none;
+    border: 3px solid #0074ff96;
+}
