+++ src/SaveTodoButton/SaveTodoButton.css
@@ -1,5 +1,6 @@
 .SaveTodoButton button {
     box-shadow: rgba(0, 0, 0, 0.6) 0 1px 5px;
+    margin: 20px 0;
 }
 
 .SaveTodoButton button:hover {
