--- src/TodoLabel.jsx
@@ -1,10 +1,0 @@
-import EmojiButton from "./EmojiButton";
-
-export default function TodoLabel({ list, action }) {
-    return (
-        <div>
-            <label htmlFor="todoText">ToDo</label>
-            <EmojiButton />
-        </div>
-    )
-};
